Cleanup in jobscontroller. HTTP Response is now formatted correctly.

+++ src/main/java/p7gruppe/p7/offloading/converters/JobConverter.java
@@ -12,7 +12,7 @@ class JobConverter implements Converter<String, Job> {
     public Job convert(String source) {
         Job job = new Job();
         // TODO: 16/11/2020 Parse this object correctly! - Philip
-        job.setId((int)System.currentTimeMillis());
+        job.setId(System.currentTimeMillis());
         job.setJobpath(source);
         job.setName(source);
         return new Job();
